processing-service/requirements.txt
pika>=1.3.2
redis>=5.0.1
python-dotenv>=1.0.0
tenacity>=8.2.3
